Rework windows build
According to upstream https://github.com/lballabio/QuantLib/issues/1552
build_shared_libs is not supported on Windows, so build a static
QuantLib library but use the dynamic runtime so the library can be used
with python.

Furthermore use an out-of-source build directory as done in the upstream
documentation https://www.quantlib.org/install/cmake.shtml this results
in a correctly generated quantlib.hpp file so there is no need to
replace it in the quantlib-python recipe.

Furthermore I enabled building the test suite on Windows and running the
tests in the test step similar to on Unix.

I also used ninja as the build tool on Windows because it was
significantly faster on my local machine and allowed faster testing.